refactor(app): move hook to local resources
