+++ app/src/organisms/Desktop/Devices/ProtocolRun/SetupCamera/index.tsx
@@ -25,7 +25,7 @@ import { useRobotStorageInfo } from '/app/resources/health/useIsImageStorageLow'
 import styles from './setupcamera.module.css'
 
 import type { CameraData } from '@opentrons/api-client'
-import type { UseCameraUsageSettingsResult } from '/app/organisms/Desktop/Devices/RobotSettings/RobotSettingsCamera/hooks/useCameraUsageSettings'
+import type { UseCameraUsageSettingsResult } from '/app/local-resources/images/hooks/useCameraUsageSettings'
 import type { State } from '/app/redux/types'
 
 export interface SetupCameraProps {
