+++ app/src/organisms/ODD/CameraSettings/UsagePreferencesSettings.tsx
@@ -6,8 +6,7 @@ import { OnOffToggle } from '/app/organisms/ODD/RobotSettingsDashboard'
 
 import styles from './preferences.module.css'
 
-// eslint-disable-next-line opentrons/no-imports-across-applications -- For active dev only
-import type { UseCameraUsageSettingsResult } from '/app/organisms/Desktop/Devices/RobotSettings/RobotSettingsCamera/hooks/useCameraUsageSettings'
+import type { UseCameraUsageSettingsResult } from '/app/local-resources/images/hooks/useCameraUsageSettings'
 
 export interface UsagePreferencesSettingsProps {
   toggleLiveVideoEnabled: UseCameraUsageSettingsResult['toggleLiveVideoEnabled']
